key_point座標描画用メソッド作成

+++ HeadDirectionWpf/OpenposeOutput.cs
@@ -23,6 +23,6 @@ namespace HeadDirectionWpf
 
     class People
     {
-        public double[] Pose_keypoints_2d { get; set; }
+        public float[] Pose_keypoints_2d { get; set; }
     }
 }
